backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
numpy==1.24.3
numexpr==2.8.4
pydantic==2.4.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
redis==5.0.1
aioredis==2.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2